Split RenameTests and CrossLanguageRenameTests to separate files
