+++ Tests/SourceKitLSPTests/RenameTests.swift
@@ -10,190 +10,9 @@
 //
 //===----------------------------------------------------------------------===//
 
-import LSPTestSupport
 import LanguageServerProtocol
-import SKSupport
 import SKTestSupport
-import SourceKitLSP
 import XCTest
-
-import enum PackageLoading.Platform
-
-private func apply(edits: [TextEdit], to source: String) -> String {
-  var lineTable = LineTable(source)
-  let edits = edits.sorted(by: { $0.range.lowerBound < $1.range.lowerBound })
-  for edit in edits.reversed() {
-    lineTable.replace(
-      fromLine: edit.range.lowerBound.line,
-      utf16Offset: edit.range.lowerBound.utf16index,
-      toLine: edit.range.upperBound.line,
-      utf16Offset: edit.range.upperBound.utf16index,
-      with: edit.newText
-    )
-  }
-  return lineTable.content
-}
-
-/// Perform a rename request at every location marker in `markedSource`, renaming it to `newName`.
-/// Test that applying the edits returned from the requests always result in `expected`.
-private func assertSingleFileRename(
-  _ markedSource: String,
-  language: Language? = nil,
-  newName: String,
-  expectedPrepareRenamePlaceholder: String,
-  expected: String,
-  testName: String = #function,
-  file: StaticString = #file,
-  line: UInt = #line
-) async throws {
-  try await SkipUnless.sourcekitdSupportsRename()
-  let testClient = try await TestSourceKitLSPClient()
-  let uri = DocumentURI.for(language ?? .swift, testName: testName)
-  let positions = testClient.openDocument(markedSource, uri: uri, language: language)
-  guard !positions.allMarkers.isEmpty else {
-    XCTFail("Test case did not contain any markers at which to invoke the rename", file: file, line: line)
-    return
-  }
-  for marker in positions.allMarkers {
-    let prepareRenameResponse = try await testClient.send(
-      PrepareRenameRequest(textDocument: TextDocumentIdentifier(uri), position: positions[marker])
-    )
-    XCTAssertEqual(
-      prepareRenameResponse?.placeholder,
-      expectedPrepareRenamePlaceholder,
-      "Prepare rename placeholder does not match while performing rename at \(marker)",
-      file: file,
-      line: line
-    )
-
-    let response = try await testClient.send(
-      RenameRequest(
-        textDocument: TextDocumentIdentifier(uri),
-        position: positions[marker],
-        newName: newName
-      )
-    )
-    let edits = try XCTUnwrap(response?.changes?[uri], "while performing rename at \(marker)", file: file, line: line)
-    let source = extractMarkers(markedSource).textWithoutMarkers
-    let renamed = apply(edits: edits, to: source)
-    XCTAssertEqual(renamed, expected, "while performing rename at \(marker)", file: file, line: line)
-  }
-}
-
-/// Assert that applying changes to `originalFiles` results in `expected`.
-///
-/// Upon failure, `message` is added to the XCTest failure messages to provide context which rename failed.
-private func assertRenamedSourceMatches(
-  originalFiles: [RelativeFileLocation: String],
-  changes: [DocumentURI: [TextEdit]],
-  expected: [RelativeFileLocation: String],
-  in ws: MultiFileTestWorkspace,
-  message: String,
-  testName: String = #function,
-  file: StaticString = #file,
-  line: UInt = #line
-) throws {
-  for (expectedFileLocation, expectedRenamed) in expected {
-    let originalMarkedSource = try XCTUnwrap(
-      originalFiles[expectedFileLocation],
-      "No original source for \(expectedFileLocation.fileName) specified; \(message)",
-      file: file,
-      line: line
-    )
-    let originalSource = extractMarkers(originalMarkedSource).textWithoutMarkers
-    let edits = changes[try ws.uri(for: expectedFileLocation.fileName)] ?? []
-    let renamed = apply(edits: edits, to: originalSource)
-    XCTAssertEqual(
-      renamed,
-      expectedRenamed,
-      "applying edits did not match expected renamed source for \(expectedFileLocation.fileName); \(message)",
-      file: file,
-      line: line
-    )
-  }
-}
-
-/// Perform a rename request at every location marker except 0️⃣ in `files`, renaming it to `newName`. The location
-/// marker 0️⃣ is intended to be used as an anchor for `preRenameActions`.
-///
-/// Test that applying the edits returned from the requests always result in `expected`.
-///
-/// `preRenameActions` is executed after opening the workspace but before performing the rename. This allows a workspace
-/// to be placed in a state where there are in-memory changes that haven't been written to disk yet.
-private func assertMultiFileRename(
-  files: [RelativeFileLocation: String],
-  headerFileLanguage: Language? = nil,
-  newName: String,
-  expectedPrepareRenamePlaceholder: String,
-  expected: [RelativeFileLocation: String],
-  manifest: String = SwiftPMTestWorkspace.defaultPackageManifest,
-  preRenameActions: (SwiftPMTestWorkspace) throws -> Void = { _ in },
-  testName: String = #function,
-  file: StaticString = #file,
-  line: UInt = #line
-) async throws {
-  try await SkipUnless.sourcekitdSupportsRename()
-  let ws = try await SwiftPMTestWorkspace(
-    files: files,
-    manifest: manifest,
-    build: true,
-    testName: testName
-  )
-  try preRenameActions(ws)
-  for (fileLocation, markedSource) in files.sorted(by: { $0.key.fileName < $1.key.fileName }) {
-    let markers = extractMarkers(markedSource).markers.keys.sorted().filter { $0 != "0️⃣" }
-    if markers.isEmpty {
-      continue
-    }
-    let (uri, positions) = try ws.openDocument(
-      fileLocation.fileName,
-      language: fileLocation.fileName.hasSuffix(".h") ? headerFileLanguage : nil
-    )
-    defer {
-      ws.testClient.send(DidCloseTextDocumentNotification(textDocument: TextDocumentIdentifier(uri)))
-    }
-    for marker in markers {
-      let prepareRenameResponse = try await ws.testClient.send(
-        PrepareRenameRequest(textDocument: TextDocumentIdentifier(uri), position: positions[marker])
-      )
-      XCTAssertEqual(
-        prepareRenameResponse?.placeholder,
-        expectedPrepareRenamePlaceholder,
-        "Prepare rename placeholder does not match while performing rename at \(marker)",
-        file: file,
-        line: line
-      )
-
-      let response = try await ws.testClient.send(
-        RenameRequest(textDocument: TextDocumentIdentifier(uri), position: positions[marker], newName: newName)
-      )
-      let changes = try XCTUnwrap(response?.changes, "Did not receive any edits", file: file, line: line)
-      try assertRenamedSourceMatches(
-        originalFiles: files,
-        changes: changes,
-        expected: expected,
-        in: ws,
-        message: "while performing rename at \(marker)",
-        file: file,
-        line: line
-      )
-    }
-  }
-}
-
-private let libAlibBPackageManifest = """
-  // swift-tools-version: 5.7
-
-  import PackageDescription
-
-  let package = Package(
-    name: "MyLibrary",
-    targets: [
-     .target(name: "LibA"),
-     .target(name: "LibB", dependencies: ["LibA"]),
-    ]
-  )
-  """
 
 final class RenameTests: XCTestCase {
   func testRenameVariableBaseName() async throws {
@@ -693,7 +512,19 @@ final class RenameTests: XCTestCase {
         }
         """,
       ],
-      manifest: libAlibBPackageManifest
+      manifest: """
+        // swift-tools-version: 5.7
+
+        import PackageDescription
+
+        let package = Package(
+          name: "MyLibrary",
+          targets: [
+          .target(name: "LibA"),
+          .target(name: "LibB", dependencies: ["LibA"]),
+          ]
+        )
+        """
     )
   }
 
@@ -911,729 +742,6 @@ final class RenameTests: XCTestCase {
         @end
         """,
       ]
-    )
-  }
-}
-
-final class CrossLanguageRenameTests: XCTestCase {
-  func testZeroArgCFunction() async throws {
-    try await SkipUnless.clangdSupportsIndexBasedRename()
-    try await assertMultiFileRename(
-      files: [
-        "LibA/include/LibA.h": """
-        void 1️⃣cFunc();
-        """,
-        "LibA/LibA.c": """
-        #include "LibA.h"
-
-        void 2️⃣cFunc() {}
-        """,
-        "LibB/LibB.swift": """
-        import LibA
-        public func test() {
-          3️⃣cFunc()
-        }
-        """,
-      ],
-      headerFileLanguage: .c,
-      newName: "dFunc",
-      expectedPrepareRenamePlaceholder: "cFunc",
-      expected: [
-        "LibA/include/LibA.h": """
-        void dFunc();
-        """,
-        "LibA/LibA.c": """
-        #include "LibA.h"
-
-        void dFunc() {}
-        """,
-        "LibB/LibB.swift": """
-        import LibA
-        public func test() {
-          dFunc()
-        }
-        """,
-      ],
-      manifest: libAlibBPackageManifest
-    )
-  }
-
-  func testMultiArgCFunction() async throws {
-    try await SkipUnless.clangdSupportsIndexBasedRename()
-    try await assertMultiFileRename(
-      files: [
-        "LibA/include/LibA.h": """
-        void 1️⃣cFunc(int x, int y);
-        """,
-        "LibA/LibA.c": """
-        #include "LibA.h"
-
-        void 2️⃣cFunc(int x, int y) {}
-        """,
-        "LibB/LibB.swift": """
-        import LibA
-        public func test() {
-          3️⃣cFunc(1, 2)
-        }
-        """,
-      ],
-      newName: "dFunc",
-      expectedPrepareRenamePlaceholder: "cFunc",
-      expected: [
-        "LibA/include/LibA.h": """
-        void dFunc(int x, int y);
-        """,
-        "LibA/LibA.c": """
-        #include "LibA.h"
-
-        void dFunc(int x, int y) {}
-        """,
-        "LibB/LibB.swift": """
-        import LibA
-        public func test() {
-          dFunc(1, 2)
-        }
-        """,
-      ],
-      manifest: libAlibBPackageManifest
-    )
-  }
-
-  func testCFunctionWithSwiftNameAnnotation() async throws {
-    try await SkipUnless.clangdSupportsIndexBasedRename()
-    try await assertMultiFileRename(
-      files: [
-        "LibA/include/LibA.h": """
-        void 1️⃣cFunc(int x, int y) __attribute__((swift_name("cFunc(x:y:)")));
-        """,
-        "LibA/LibA.c": """
-        #include "LibA.h"
-
-        void 2️⃣cFunc(int x, int y) {}
-        """,
-        "LibB/LibB.swift": """
-        import LibA
-        public func test() {
-          3️⃣cFunc(x: 1, y: 2)
-        }
-        """,
-      ],
-      headerFileLanguage: .c,
-      newName: "dFunc",
-      expectedPrepareRenamePlaceholder: "cFunc",
-      expected: [
-        "LibA/include/LibA.h": """
-        void dFunc(int x, int y) __attribute__((swift_name("cFunc(x:y:)")));
-        """,
-        "LibA/LibA.c": """
-        #include "LibA.h"
-
-        void dFunc(int x, int y) {}
-        """,
-        "LibB/LibB.swift": """
-        import LibA
-        public func test() {
-          cFunc(x: 1, y: 2)
-        }
-        """,
-      ],
-      manifest: libAlibBPackageManifest
-    )
-  }
-
-  func testZeroArgObjCSelector() async throws {
-    try SkipUnless.platformIsDarwin("Non-Darwin platforms don't support Objective-C")
-    try await SkipUnless.clangdSupportsIndexBasedRename()
-    try await assertMultiFileRename(
-      files: [
-        "LibA/include/LibA.h": """
-        @interface Foo
-        - (int)1️⃣performAction;
-        @end
-        """,
-        "LibA/LibA.m": """
-        #include "LibA.h"
-
-        @implementation Foo
-        - (int)performAction {
-          return [self 2️⃣performAction];
-        }
-        @end
-        """,
-        "LibB/LibB.swift": """
-        import LibA
-        public func test(foo: Foo) {
-          foo.3️⃣performAction()
-        }
-        """,
-      ],
-      headerFileLanguage: .objective_c,
-      newName: "performNewAction",
-      expectedPrepareRenamePlaceholder: "performAction",
-      expected: [
-        "LibA/include/LibA.h": """
-        @interface Foo
-        - (int)performNewAction;
-        @end
-        """,
-        "LibA/LibA.m": """
-        #include "LibA.h"
-
-        @implementation Foo
-        - (int)performNewAction {
-          return [self performNewAction];
-        }
-        @end
-        """,
-        "LibB/LibB.swift": """
-        import LibA
-        public func test(foo: Foo) {
-          foo.performNewAction()
-        }
-        """,
-      ],
-      manifest: libAlibBPackageManifest
-    )
-  }
-
-  func testZeroArgObjCClassSelector() async throws {
-    try SkipUnless.platformIsDarwin("Non-Darwin platforms don't support Objective-C")
-    try await SkipUnless.clangdSupportsIndexBasedRename()
-    try await assertMultiFileRename(
-      files: [
-        "LibA/include/LibA.h": """
-        @interface Foo
-        + (int)1️⃣performAction;
-        @end
-        """,
-        "LibA/LibA.m": """
-        #include "LibA.h"
-
-        @implementation Foo
-        + (int)performAction {
-          return [Foo 2️⃣performAction];
-        }
-        @end
-        """,
-        "LibB/LibB.swift": """
-        import LibA
-        public func test() {
-          Foo.3️⃣performAction()
-        }
-        """,
-      ],
-      headerFileLanguage: .objective_c,
-      newName: "performNewAction",
-      expectedPrepareRenamePlaceholder: "performAction",
-      expected: [
-        "LibA/include/LibA.h": """
-        @interface Foo
-        + (int)performNewAction;
-        @end
-        """,
-        "LibA/LibA.m": """
-        #include "LibA.h"
-
-        @implementation Foo
-        + (int)performNewAction {
-          return [Foo performNewAction];
-        }
-        @end
-        """,
-        "LibB/LibB.swift": """
-        import LibA
-        public func test() {
-          Foo.performNewAction()
-        }
-        """,
-      ],
-      manifest: libAlibBPackageManifest
-    )
-  }
-
-  func testOneArgObjCSelector() async throws {
-    try SkipUnless.platformIsDarwin("Non-Darwin platforms don't support Objective-C")
-    try await SkipUnless.clangdSupportsIndexBasedRename()
-    try await assertMultiFileRename(
-      files: [
-        "LibA/include/LibA.h": """
-        @interface Foo
-        - (int)1️⃣performAction:(int)action;
-        @end
-        """,
-        "LibA/LibA.m": """
-        #include "LibA.h"
-
-        @implementation Foo
-        - (int)2️⃣performAction:(int)action {
-          return [self performAction:action];
-        }
-        @end
-        """,
-        "LibB/LibB.swift": """
-        import LibA
-        public func test(foo: Foo) {
-          foo.3️⃣performAction(1)
-        }
-        """,
-      ],
-      headerFileLanguage: .objective_c,
-      newName: "performNewAction:",
-      expectedPrepareRenamePlaceholder: "performAction:",
-      expected: [
-        "LibA/include/LibA.h": """
-        @interface Foo
-        - (int)performNewAction:(int)action;
-        @end
-        """,
-        "LibA/LibA.m": """
-        #include "LibA.h"
-
-        @implementation Foo
-        - (int)performNewAction:(int)action {
-          return [self performNewAction:action];
-        }
-        @end
-        """,
-        "LibB/LibB.swift": """
-        import LibA
-        public func test(foo: Foo) {
-          foo.performNewAction(1)
-        }
-        """,
-      ],
-      manifest: libAlibBPackageManifest
-    )
-  }
-
-  func testMultiArgObjCSelector() async throws {
-    try SkipUnless.platformIsDarwin("Non-Darwin platforms don't support Objective-C")
-    try await SkipUnless.clangdSupportsIndexBasedRename()
-    try await assertMultiFileRename(
-      files: [
-        "LibA/include/LibA.h": """
-        @interface Foo
-        - (int)1️⃣performAction:(int)action with:(int)value;
-        @end
-        """,
-        "LibA/LibA.m": """
-        #include "LibA.h"
-
-        @implementation Foo
-        - (int)2️⃣performAction:(int)action with:(int)value {
-          return [self performAction:action with:value];
-        }
-        @end
-        """,
-        "LibB/LibB.swift": """
-        import LibA
-        public func test(foo: Foo) {
-          foo.3️⃣performAction(1, with: 2)
-        }
-        """,
-      ],
-      headerFileLanguage: .objective_c,
-      newName: "performNewAction:by:",
-      expectedPrepareRenamePlaceholder: "performAction:with:",
-      expected: [
-        "LibA/include/LibA.h": """
-        @interface Foo
-        - (int)performNewAction:(int)action by:(int)value;
-        @end
-        """,
-        "LibA/LibA.m": """
-        #include "LibA.h"
-
-        @implementation Foo
-        - (int)performNewAction:(int)action by:(int)value {
-          return [self performNewAction:action by:value];
-        }
-        @end
-        """,
-        "LibB/LibB.swift": """
-        import LibA
-        public func test(foo: Foo) {
-          foo.performNewAction(1, by: 2)
-        }
-        """,
-      ],
-      manifest: libAlibBPackageManifest
-    )
-  }
-
-  func testObjCSelectorWithSwiftNameAnnotation() async throws {
-    try SkipUnless.platformIsDarwin("Non-Darwin platforms don't support Objective-C")
-    try await SkipUnless.clangdSupportsIndexBasedRename()
-    try await assertMultiFileRename(
-      files: [
-        "LibA/include/LibA.h": """
-        @interface Foo
-        - (int)1️⃣performAction:(int)action withValue:(int)value __attribute__((swift_name("perform(action:with:)")));
-        @end
-        """,
-        "LibA/LibA.m": """
-        #include "LibA.h"
-
-        @implementation Foo
-        - (int)2️⃣performAction:(int)action withValue:(int)value {
-          return [self performAction:action withValue:value];
-        }
-        @end
-        """,
-        "LibB/LibB.swift": """
-        import LibA
-        public func test(foo: Foo) {
-          foo.3️⃣perform(action: 1, with: 2)
-        }
-        """,
-      ],
-      headerFileLanguage: .objective_c,
-      newName: "performNewAction:by:",
-      expectedPrepareRenamePlaceholder: "performAction:withValue:",
-      expected: [
-        "LibA/include/LibA.h": """
-        @interface Foo
-        - (int)performNewAction:(int)action by:(int)value __attribute__((swift_name("perform(action:with:)")));
-        @end
-        """,
-        "LibA/LibA.m": """
-        #include "LibA.h"
-
-        @implementation Foo
-        - (int)performNewAction:(int)action by:(int)value {
-          return [self performNewAction:action by:value];
-        }
-        @end
-        """,
-        "LibB/LibB.swift": """
-        import LibA
-        public func test(foo: Foo) {
-          foo.perform(action: 1, with: 2)
-        }
-        """,
-      ],
-      manifest: libAlibBPackageManifest
-    )
-  }
-
-  func testObjCClass() async throws {
-    try SkipUnless.platformIsDarwin("Non-Darwin platforms don't support Objective-C")
-    try await SkipUnless.clangdSupportsIndexBasedRename()
-    try await assertMultiFileRename(
-      files: [
-        "LibA/include/LibA.h": """
-        @interface 1️⃣Foo
-        @end
-        """,
-        "LibA/LibA.m": """
-        #include "LibA.h"
-
-        @implementation 2️⃣Foo
-        @end
-        """,
-        "LibB/LibB.swift": """
-        import LibA
-        public func test(foo: 3️⃣Foo) {
-        }
-        """,
-      ],
-      headerFileLanguage: .objective_c,
-      newName: "Bar",
-      expectedPrepareRenamePlaceholder: "Foo",
-      expected: [
-        "LibA/include/LibA.h": """
-        @interface Bar
-        @end
-        """,
-        "LibA/LibA.m": """
-        #include "LibA.h"
-
-        @implementation Bar
-        @end
-        """,
-        "LibB/LibB.swift": """
-        import LibA
-        public func test(foo: Bar) {
-        }
-        """,
-      ],
-      manifest: libAlibBPackageManifest
-    )
-  }
-
-  func testObjCClassWithSwiftNameAnnotation() async throws {
-    try SkipUnless.platformIsDarwin("Non-Darwin platforms don't support Objective-C")
-    try await SkipUnless.clangdSupportsIndexBasedRename()
-    try await assertMultiFileRename(
-      files: [
-        "LibA/include/LibA.h": """
-        __attribute__((swift_name("Foo")))
-        @interface 1️⃣AHFoo
-        @end
-        """,
-        "LibA/LibA.m": """
-        #include "LibA.h"
-
-        @implementation 2️⃣AHFoo
-        @end
-        """,
-        "LibB/LibB.swift": """
-        import LibA
-        public func test(foo: 3️⃣Foo) {
-        }
-        """,
-      ],
-      headerFileLanguage: .objective_c,
-      newName: "AHBar",
-      expectedPrepareRenamePlaceholder: "AHFoo",
-      expected: [
-        "LibA/include/LibA.h": """
-        __attribute__((swift_name("Foo")))
-        @interface AHBar
-        @end
-        """,
-        "LibA/LibA.m": """
-        #include "LibA.h"
-
-        @implementation AHBar
-        @end
-        """,
-        "LibB/LibB.swift": """
-        import LibA
-        public func test(foo: Foo) {
-        }
-        """,
-      ],
-      manifest: libAlibBPackageManifest
-    )
-  }
-
-  func testCppMethod() async throws {
-    try await SkipUnless.clangdSupportsIndexBasedRename()
-    try await assertMultiFileRename(
-      files: [
-        "LibA/include/LibA.h": """
-        struct Foo {
-          void 1️⃣doStuff() const;
-        };
-        """,
-        "LibA/LibA.cpp": """
-        #include "LibA.h"
-
-        void Foo::2️⃣doStuff() const {};
-        """,
-        "LibB/LibB.swift": """
-        import LibA
-        public func test(foo: Foo) {
-          foo.3️⃣doStuff()
-        }
-        """,
-      ],
-      headerFileLanguage: .cpp,
-      newName: "doCoolStuff",
-      expectedPrepareRenamePlaceholder: "doStuff",
-      expected: [
-        "LibA/include/LibA.h": """
-        struct Foo {
-          void doCoolStuff() const;
-        };
-        """,
-        "LibA/LibA.cpp": """
-        #include "LibA.h"
-
-        void Foo::doCoolStuff() const {};
-        """,
-        "LibB/LibB.swift": """
-        import LibA
-        public func test(foo: Foo) {
-          foo.doCoolStuff()
-        }
-        """,
-      ],
-      manifest: """
-        // swift-tools-version: 5.7
-
-        import PackageDescription
-
-        let package = Package(
-          name: "MyLibrary",
-          targets: [
-           .target(name: "LibA"),
-           .target(name: "LibB", dependencies: ["LibA"], swiftSettings: [.unsafeFlags(["-cxx-interoperability-mode=default"])]),
-          ]
-        )
-        """
-    )
-  }
-
-  func testCppMethodWithSwiftName() async throws {
-    try await SkipUnless.clangdSupportsIndexBasedRename()
-    try await assertMultiFileRename(
-      files: [
-        "LibA/include/LibA.h": """
-        struct Foo {
-          void 1️⃣doStuff(int x) const __attribute__((swift_name("do(stuff:)")));
-        };
-        """,
-        "LibA/LibA.cpp": """
-        #include "LibA.h"
-
-        void Foo::2️⃣doStuff(int x) const {};
-        """,
-        "LibB/LibB.swift": """
-        import LibA
-        public func test(foo: Foo) {
-          foo.3️⃣do(stuff: 1)
-        }
-        """,
-      ],
-      headerFileLanguage: .cpp,
-      newName: "doCoolStuff",
-      expectedPrepareRenamePlaceholder: "doStuff",
-      expected: [
-        "LibA/include/LibA.h": """
-        struct Foo {
-          void doCoolStuff(int x) const __attribute__((swift_name("do(stuff:)")));
-        };
-        """,
-        "LibA/LibA.cpp": """
-        #include "LibA.h"
-
-        void Foo::doCoolStuff(int x) const {};
-        """,
-        "LibB/LibB.swift": """
-        import LibA
-        public func test(foo: Foo) {
-          foo.do(stuff: 1)
-        }
-        """,
-      ],
-      manifest: """
-        // swift-tools-version: 5.7
-
-        import PackageDescription
-
-        let package = Package(
-          name: "MyLibrary",
-          targets: [
-           .target(name: "LibA"),
-           .target(name: "LibB", dependencies: ["LibA"], swiftSettings: [.unsafeFlags(["-cxx-interoperability-mode=default"])]),
-          ]
-        )
-        """
-    )
-  }
-
-  func testCppMethodInObjCpp() async throws {
-    try await SkipUnless.clangdSupportsIndexBasedRename()
-    try await assertMultiFileRename(
-      files: [
-        "LibA/include/LibA.h": """
-        struct Foo {
-          void 1️⃣doStuff() const;
-        };
-        """,
-        "LibA/LibA.mm": """
-        #include "LibA.h"
-
-        void Foo::2️⃣doStuff() const {};
-        """,
-        "LibB/LibB.swift": """
-        import LibA
-        public func test(foo: Foo) {
-          foo.3️⃣doStuff()
-        }
-        """,
-      ],
-      headerFileLanguage: .objective_cpp,
-      newName: "doCoolStuff",
-      expectedPrepareRenamePlaceholder: "doStuff",
-      expected: [
-        "LibA/include/LibA.h": """
-        struct Foo {
-          void doCoolStuff() const;
-        };
-        """,
-        "LibA/LibA.mm": """
-        #include "LibA.h"
-
-        void Foo::doCoolStuff() const {};
-        """,
-        "LibB/LibB.swift": """
-        import LibA
-        public func test(foo: Foo) {
-          foo.doCoolStuff()
-        }
-        """,
-      ],
-      manifest: """
-        // swift-tools-version: 5.7
-
-        import PackageDescription
-
-        let package = Package(
-          name: "MyLibrary",
-          targets: [
-           .target(name: "LibA"),
-           .target(name: "LibB", dependencies: ["LibA"], swiftSettings: [.unsafeFlags(["-cxx-interoperability-mode=default"])]),
-          ]
-        )
-        """
-    )
-  }
-
-  func testZeroArgObjCClassSelectorInObjCpp() async throws {
-    try SkipUnless.platformIsDarwin("Non-Darwin platforms don't support Objective-C")
-    try await SkipUnless.clangdSupportsIndexBasedRename()
-    try await assertMultiFileRename(
-      files: [
-        "LibA/include/LibA.h": """
-        @interface Foo
-        + (int)1️⃣performAction;
-        @end
-        """,
-        "LibA/LibA.mm": """
-        #include "LibA.h"
-
-        @implementation Foo
-        + (int)performAction {
-          return [Foo 2️⃣performAction];
-        }
-        @end
-        """,
-        "LibB/LibB.swift": """
-        import LibA
-        public func test() {
-          Foo.3️⃣performAction()
-        }
-        """,
-      ],
-      headerFileLanguage: .objective_cpp,
-      newName: "performNewAction",
-      expectedPrepareRenamePlaceholder: "performAction",
-      expected: [
-        "LibA/include/LibA.h": """
-        @interface Foo
-        + (int)performNewAction;
-        @end
-        """,
-        "LibA/LibA.mm": """
-        #include "LibA.h"
-
-        @implementation Foo
-        + (int)performNewAction {
-          return [Foo performNewAction];
-        }
-        @end
-        """,
-        "LibB/LibB.swift": """
-        import LibA
-        public func test() {
-          Foo.performNewAction()
-        }
-        """,
-      ],
-      manifest: libAlibBPackageManifest
     )
   }
 }
